Restore initialization list, this time without using uninitialized member variables. Change a size_t to a uint32_t to match the type used in the folly PC queue

+++ dune-artdaq/Generators/Felix/TriggerPrimitive/TriggerPrimitiveFinder.h
@@ -29,7 +29,7 @@ public:
         std::vector<TriggerPrimitive> triggerPrimitives;
     };
 
-    TriggerPrimitiveFinder(size_t qsize, size_t timeWindowNumMessages, size_t nthreads);
+    TriggerPrimitiveFinder(uint32_t qsize, size_t timeWindowNumMessages, size_t nthreads);
   
     ~TriggerPrimitiveFinder();
